Writes a test program for the passive Spindex. Adjusts the outtake class to not require a kicker parameter.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Subsystems/Spindex.java
@@ -324,7 +324,7 @@ public class Spindex {
         outtakeMode = outtake;
     }
 
-    public void setPower(int power){
+    public void setPower(double power){
         spindexMotor.setPower(power);
     }
